+++ Fedoraware/Fedoraware-TF2/src/Features/Resolver/Resolver.cpp
@@ -99,7 +99,7 @@ float PResolver::GetRealPitch(const float flPitch){
 
 void PResolver::SetAngles(const Vec3 vAngles, CBaseEntity* pEntity){
 	if (CTFPlayerAnimState* pAnimState = pEntity->GetAnimState()){
-		float* flPitch = reinterpret_cast<float*>(reinterpret_cast<DWORD>(pEntity) + g_NetVars.get_offset("DT_TFPlayer", "tfnonlocaldata", "m_angEyeAngles[0]"));
+		float* flPitch = reinterpret_cast<float*>(reinterpret_cast<uintptr_t>(pEntity) + g_NetVars.get_offset("DT_TFPlayer", "tfnonlocaldata", "m_angEyeAngles[0]"));
 
 		*flPitch = vAngles.x;
 		pAnimState->m_flCurrentFeetYaw = vAngles.y;
